backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Firebase & Google Cloud
firebase-admin==6.4.0
google-cloud-storage==2.14.0
google-cloud-firestore==2.14.0
google-cloud-logging==3.9.0
google-cloud-secret-manager==2.18.0

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0

# Testing (optional)
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
httpx==0.25.2
